rewrite to use makefile instead of scripts
