--- backend/java/.gitignore
@@ -1,5 +1,0 @@
-*~
-.DS_Store
-data
-/target/
-.idea/*
