--- backend/java/src/main/java/org/unipept/storage/.gitignore
@@ -1,4 +1,0 @@
-/Database.class
-/PeptideLoaderData.class
-/TaxonInvalidatorData.class
-/TaxonLoaderData.class
