--- backend/java/src/main/java/org/unipept/tools/.gitignore
@@ -1,1 +1,0 @@
-/ProgressWriter.class
