--- backend/java/src/main/java/org/unipept/tools/commandline/.gitignore
@@ -1,3 +1,0 @@
-/PeptideLoader.class
-/TaxonInvalidator.class
-/TaxonLoader.class
