--- backend/java/src/main/java/org/unipept/xml/.gitignore
@@ -1,19 +1,0 @@
-/UniprotDbRef.class
-/UniprotECRef.class
-/UniprotEntry$Pair.class
-/UniprotEntry.class
-/UniprotGORef.class
-/UniprotHandler$1.class
-/UniprotHandler$10.class
-/UniprotHandler$11.class
-/UniprotHandler$2.class
-/UniprotHandler$3.class
-/UniprotHandler$4.class
-/UniprotHandler$5.class
-/UniprotHandler$6.class
-/UniprotHandler$7.class
-/UniprotHandler$8.class
-/UniprotHandler$9.class
-/UniprotHandler$EndTagWorker.class
-/UniprotHandler$StartTagWorker.class
-/UniprotHandler.class
