--- backend/scripts/compile.sh
@@ -1,10 +1,0 @@
-#!/bin/bash
-# compiles all java files
-
-currentdir=`pwd`
-
-cd "backend/java"
-mvn clean
-mvn package
-
-cd "${currentdir}"
