--- backend/scripts/invalidateTaxon.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-backend/java/bin/taxon_invalidator
-backend/java/bin/peptide_loader lineages
