+++ backend/src/main/java/org/unipept/xml/UniprotGORef.java
